moved fortran stuff back to state
